refactor: gooey_builder_tools.py
